+++ src/containers/Prefs/PrefsGeneral.js
@@ -50,7 +50,7 @@ class PrefsGeneral extends React.Component {
 
   componentDidMount(){
 
-    service.registerListener(this);
+    //service.registerListener(this);
     service.api.readConfKey('prefs').then((value)=>{
       this.setState({prefs: value });
 
@@ -72,7 +72,7 @@ class PrefsGeneral extends React.Component {
   }
 
   componentWillUnmount(){
-    service.unregisterListener(this);
+   // service.unregisterListener(this);
   }
 
   handleFolderSelected(folder){