+++ src/containers/Sidebar.js
@@ -1,17 +1,17 @@
-import React from 'react';
-import { withStyles } from '@material-ui/core/styles';
-import Divider                  from '@material-ui/core/Divider';
-import List                     from '@material-ui/core/List';
-import ListSubheader            from '@material-ui/core/ListSubheader';
-import ListItem                 from '@material-ui/core/ListItem';
-import ListItemIcon             from '@material-ui/core/ListItemIcon';
-import ListItemText             from '@material-ui/core/ListItemText';
-import ListItemSecondaryAction  from '@material-ui/core/ListItemSecondaryAction';
-import Box                      from '@material-ui/core/Box';
-import Collapse                 from '@material-ui/core/Collapse';
-import ExpandLess               from '@material-ui/icons/ExpandLess';
-import ExpandMore               from '@material-ui/icons/ExpandMore';
-import IconButton               from '@material-ui/core/IconButton';
+import React                   from 'react';
+import { withStyles }          from '@material-ui/core/styles';
+import Divider                 from '@material-ui/core/Divider';
+import List                    from '@material-ui/core/List';
+import ListSubheader           from '@material-ui/core/ListSubheader';
+import ListItem                from '@material-ui/core/ListItem';
+import ListItemIcon            from '@material-ui/core/ListItemIcon';
+import ListItemText            from '@material-ui/core/ListItemText';
+import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
+import Box                     from '@material-ui/core/Box';
+import Collapse                from '@material-ui/core/Collapse';
+import ExpandLess              from '@material-ui/icons/ExpandLess';
+import ExpandMore              from '@material-ui/icons/ExpandMore';
+import IconButton              from '@material-ui/core/IconButton';
 
 const useStyles = theme => ({
   nested: {